+++ persistence/kvstore/kvstore.go
@@ -1,22 +1,23 @@
 package kvstore
 
 import (
+	"errors"
 	"log"
 
+	"github.com/celestiaorg/smt"
 	badger "github.com/dgraph-io/badger/v3"
 )
 
-// CLEANUP: move this structure to a shared module
 type KVStore interface {
+	smt.MapStore // Get, Set, Delete
+
 	// Lifecycle methods
 	Stop() error
 
 	// Accessors
 	// TODO: Add a proper iterator interface
-	Put(key []byte, value []byte) error
-	Get(key []byte) ([]byte, error)
 	// TODO: Add pagination for `GetAll`
-	GetAll(prefixKey []byte, descending bool) ([][]byte, error)
+	GetAll(prefixKey []byte, descending bool) (keys [][]byte, values [][]byte, err error)
 	Exists(key []byte) (bool, error)
 	ClearAll() error
 }
@@ -26,28 +27,34 @@ const (
 )
 
 var _ KVStore = &badgerKVStore{}
+var _ smt.MapStore = &badgerKVStore{}
+
+var (
+	ErrKVStoreExists    = errors.New("kvstore already exists")
+	ErrKVStoreNotExists = errors.New("kvstore does not exist")
+)
 
 type badgerKVStore struct {
 	db *badger.DB
 }
 
 func NewKVStore(path string) (KVStore, error) {
-	db, err := badger.Open(badger.DefaultOptions(path))
+	db, err := badger.Open(badgerOptions(path))
 	if err != nil {
 		return nil, err
 	}
-	return badgerKVStore{db: db}, nil
+	return &badgerKVStore{db: db}, nil
 }
 
 func NewMemKVStore() KVStore {
-	db, err := badger.Open(badger.DefaultOptions("").WithInMemory(true))
+	db, err := badger.Open(badgerOptions("").WithInMemory(true))
 	if err != nil {
 		log.Fatal(err)
 	}
-	return badgerKVStore{db: db}
+	return &badgerKVStore{db: db}
 }
 
-func (store badgerKVStore) Put(key []byte, value []byte) error {
+func (store *badgerKVStore) Set(key, value []byte) error {
 	tx := store.db.NewTransaction(true)
 	defer tx.Discard()
 
@@ -59,7 +66,7 @@ func (store badgerKVStore) Put(key []byte, value []byte) error {
 	return tx.Commit()
 }
 
-func (store badgerKVStore) Get(key []byte) ([]byte, error) {
+func (store *badgerKVStore) Get(key []byte) ([]byte, error) {
 	tx := store.db.NewTransaction(false)
 	defer tx.Discard()
 
@@ -80,8 +87,16 @@ func (store badgerKVStore) Get(key []byte) ([]byte, error) {
 	return value, nil
 }
 
-func (store badgerKVStore) GetAll(prefix []byte, descending bool) (values [][]byte, err error) {
+func (store *badgerKVStore) Delete(key []byte) error {
+	tx := store.db.NewTransaction(true)
+	defer tx.Discard()
+
+	return tx.Delete(key)
+}
+
+func (store *badgerKVStore) GetAll(prefix []byte, descending bool) (keys [][]byte, values [][]byte, err error) {
 	// INVESTIGATE: research `badger.views` for further improvements and optimizations
+	// Reference https://pkg.go.dev/github.com/dgraph-io/badger#readme-prefix-scans
 	txn := store.db.NewTransaction(false)
 	defer txn.Discard()
 
@@ -94,11 +109,15 @@ func (store badgerKVStore) GetAll(prefix []byte, descending bool) (values [][]by
 	it := txn.NewIterator(opt)
 	defer it.Close()
 
+	keys = make([][]byte, 0)
+	values = make([][]byte, 0)
+
 	for it.Seek(prefix); it.Valid(); it.Next() {
 		item := it.Item()
 		err = item.Value(func(v []byte) error {
 			b := make([]byte, len(v))
 			copy(b, v)
+			keys = append(keys, item.Key())
 			values = append(values, b)
 			return nil
 		})
@@ -109,7 +128,7 @@ func (store badgerKVStore) GetAll(prefix []byte, descending bool) (values [][]by
 	return
 }
 
-func (store badgerKVStore) Exists(key []byte) (bool, error) {
+func (store *badgerKVStore) Exists(key []byte) (bool, error) {
 	val, err := store.Get(key)
 	if err != nil {
 		return false, err
@@ -117,11 +136,11 @@ func (store badgerKVStore) Exists(key []byte) (bool, error) {
 	return val != nil, nil
 }
 
-func (store badgerKVStore) ClearAll() error {
+func (store *badgerKVStore) ClearAll() error {
 	return store.db.DropAll()
 }
 
-func (store badgerKVStore) Stop() error {
+func (store *badgerKVStore) Stop() error {
 	return store.db.Close()
 }
 
@@ -140,4 +159,11 @@ func prefixEndBytes(prefix []byte) []byte {
 	copy(end, prefix)
 	end[len(end)-1]++
 	return end
+}
+
+// TODO: Propagate persistence configurations to badger
+func badgerOptions(path string) badger.Options {
+	opts := badger.DefaultOptions(path)
+	opts.Logger = nil // disable badger's logger since it's very noisy
+	return opts
 }