+++ persistence/test/fisherman_test.go
@@ -26,6 +26,13 @@ func TestGetSetFishermanStakeAmount(t *testing.T) {
 	getTestGetSetStakeAmountTest(t, db, createAndInsertDefaultTestFisherman, db.GetFishermanStakeAmount, db.SetFishermanStakeAmount, 1)
 }
 
+func TestGetFishermanUpdatedAtHeight(t *testing.T) {
+	getFishermanUpdatedFunc := func(db *persistence.PostgresContext, height int64) ([]*types.Actor, error) {
+		return db.GetActorsUpdated(types.FishermanActor, height)
+	}
+	getAllActorsUpdatedAtHeightTest(t, createAndInsertDefaultTestFisherman, getFishermanUpdatedFunc, 1)
+}
+
 func TestInsertFishermanAndExists(t *testing.T) {
 	db := NewTestPostgresContext(t, 0)